grammatical error correction in variable names
